separate MiniDataset type per component; ts fixes
