+++ src/components/src/side-panel/common/dataset-info.tsx
@@ -22,9 +22,17 @@ import React from 'react';
 import styled from 'styled-components';
 import {format} from 'd3-format';
 import {FormattedMessage} from '@kepler.gl/localization';
-import {DatasetInfoProps} from './types';
+import {DataContainerInterface} from '@kepler.gl/utils';
 
 const numFormat = format(',');
+
+type MiniDataset = {
+  dataContainer: DataContainerInterface;
+};
+
+export type DatasetInfoProps = {
+  dataset: MiniDataset;
+};
 
 const StyledDataRowCount = styled.div`
   font-size: 11px;
