+++ src/components/src/side-panel/common/source-data-catalog.tsx
@@ -21,14 +21,37 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import {openDeleteModal, VisStateActions, ActionHandler} from '@kepler.gl/actions';
+import {DataContainerInterface} from '@kepler.gl/utils';
+import {RGBColor} from '@kepler.gl/types';
+
 import {SidePanelSection} from '../../common/styled-components';
 import DatasetTitleFactory from './dataset-title';
 import DatasetInfoFactory from './dataset-info';
-import {SourceDataCatalogProps} from './types';
 
 const SourceDataCatalogWrapper = styled.div`
   transition: ${props => props.theme.transition};
 `;
+
+type MiniDataset = {
+  id: string;
+  color: RGBColor;
+  label?: string;
+  dataContainer: DataContainerInterface;
+};
+
+type MiniDatasets = {
+  [key: string]: MiniDataset;
+};
+
+export type SourceDataCatalogProps = {
+  datasets: MiniDatasets;
+  showDeleteDataset?: boolean;
+  onTitleClick?: () => void;
+  showDatasetTable?: ActionHandler<typeof VisStateActions.showDatasetTable>;
+  updateTableColor: ActionHandler<typeof VisStateActions.updateTableColor>;
+  removeDataset?: ActionHandler<typeof openDeleteModal>;
+};
 
 SourceDataCatalogFactory.deps = [DatasetTitleFactory, DatasetInfoFactory];
