+++ src/components/src/side-panel/common/types.ts
@@ -1,13 +1,7 @@
 import React from 'react';
 import {openDeleteModal, VisStateActions, ActionHandler} from '@kepler.gl/actions';
 import {RGBColor} from '@kepler.gl/types';
-import KeplerTable, {Datasets} from '@kepler.gl/table';
-
-type MiniDataset = {
-  id: string;
-  color: RGBColor;
-  label?: string;
-};
+import KeplerTable from '@kepler.gl/table';
 
 export type PanelMeta = {
   id: string;
@@ -16,21 +10,9 @@ export type PanelMeta = {
   onClick: null;
 };
 
-export type DatasetInfoProps = {
-  dataset: KeplerTable;
-};
-
 export type UpdateTableColorTypes = {
   id: string;
   children: React.ReactNode;
-};
-
-export type DatasetTagProps = {
-  id?: string;
-  dataset: MiniDataset;
-  updateTableColor?: ActionHandler<typeof VisStateActions.updateTableColor>;
-  onClick?: React.MouseEventHandler<HTMLDivElement>;
-  onClickSquare?: React.MouseEventHandler<HTMLDivElement>;
 };
 
 export type ShowDataTableProps = {
@@ -45,24 +27,6 @@ export type RemoveDatasetProps = {
 
 export type StyledDatasetTitleProps = {
   clickable: boolean;
-};
-
-export type DatasetTitleProps = {
-  dataset: KeplerTable;
-  showDeleteDataset: boolean;
-  onTitleClick?: () => void;
-  showDatasetTable?: ActionHandler<typeof VisStateActions.showDatasetTable>;
-  updateTableColor: ActionHandler<typeof VisStateActions.updateTableColor>;
-  removeDataset?: ActionHandler<typeof openDeleteModal>;
-};
-
-export type SourceDataCatalogProps = {
-  datasets: Datasets;
-  showDeleteDataset?: boolean;
-  onTitleClick?: () => void;
-  showDatasetTable?: ActionHandler<typeof VisStateActions.showDatasetTable>;
-  updateTableColor: ActionHandler<typeof VisStateActions.updateTableColor>;
-  removeDataset?: ActionHandler<typeof openDeleteModal>;
 };
 
 export type DatasetItemProps = {
